+++ mkdocs.yml
@@ -1,3 +1,3 @@
 site_name: My Docs
 theme:
-  name: mkdocs
+  name: material
